+++ warehub/warehub/model/ProductFactory.cs
@@ -6,12 +6,12 @@ namespace warehub.model
 {
     public static class ProductFactory
     {
-        public static Product CreateProduct(string name, int price)
+        public static Product CreateProduct(string name, decimal price)
         {
             return new Product(name, price);
         }
 
-        public static Product CreateProduct(Guid id, string name, int price)
+        public static Product CreateProduct(Guid id, string name, decimal price)
         {
             return new Product(id, name, price);
         }
